[Chore] #68 Enum 폴더 생성하여 파일을 만들고 enum을 사용한 코드 이동
추가적으로 변수명도 ~Style 형태로 통일하였다
