+++ Milestone/Milestone/Global/Enum/ButtonStateStyle.swift
@@ -1,0 +1,15 @@
+//
+//  ButtonStateStyle.swift
+//  Milestone
+//
+//  Created by 서은수 on 2023/08/20.
+//
+
+import UIKit
+
+/// 버튼의 상태를 나타내는 enum
+enum ButtonState {
+    case original
+    case press
+    case disabled
+}
