+++ Milestone/Milestone/Global/Enum/DayStyle.swift
@@ -1,0 +1,19 @@
+//
+//  enum DayStyle.swift
+//  Milestone
+//
+//  Created by 서은수 on 2023/08/20.
+//
+
+import UIKit
+
+/// 요일 데이터를 나타내는 enum
+enum Days: String {
+    case MONDAY = "월"
+    case TUEDAY = "화"
+    case WEDDAY = "수"
+    case THUDAY = "목"
+    case FRIDAY = "금"
+    case SATDAY = "토"
+    case SUNDAY = "일"
+}
