+++ Milestone/Milestone/Global/Enum/IndexTextStyle.swift
@@ -1,0 +1,15 @@
+//
+//  IndexTextStyle.swift
+//  Milestone
+//
+//  Created by 서은수 on 2023/08/20.
+//
+
+import UIKit
+
+enum IndexTextStyle: String {
+    case first = "좋았던 점은 무엇인가요?"
+    case second = "아쉬웠던 점이나, 부족했던 점은 무엇인가요?"
+    case third = "배운 점은 무엇인가요?"
+    case fourth = "목표를 통해 뭘 얻고자 하셨나요?"
+}
