+++ Milestone/Milestone/Global/Enum/PointTextStyle.swift
@@ -1,0 +1,16 @@
+//
+//  PointTextStyle.swift
+//  Milestone
+//
+//  Created by 서은수 on 2023/08/20.
+//
+
+import UIKit
+
+enum PointTextStyle: String {
+    case lowest = "별로예요"
+    case lower = "아쉬워요"
+    case middle = "그저 그랬어요"
+    case higher = "만족해요"
+    case highest = "완전 만족해요"
+}
