+++ Milestone/Milestone/Global/Enum/UserDefaultsKeyStyle.swift
@@ -1,0 +1,13 @@
+//
+//  UserDefaultsKeyStyle.swift
+//  Milestone
+//
+//  Created by 서은수 on 2023/08/20.
+//
+
+import UIKit
+
+/// UserDefaults에 사용되는 key 값 모음
+enum UserDefaultsKeyStyle: String {
+    case couchMark = "showCouchMark"
+}
