+++ Milestone/Milestone/Global/Enum/WriteRetrospectStyle.swift
@@ -1,0 +1,13 @@
+//
+//  WriteRetrospectStyle.swift
+//  Milestone
+//
+//  Created by 서은수 on 2023/08/20.
+//
+
+import UIKit
+
+enum WriteRetrospectStyle {
+    case free
+    case guide
+}
