+++ Milestone/Milestone/Screens/FillBox/View/DetailParentViewController.swift
@@ -12,12 +12,6 @@ import RxSwift
 import SnapKit
 import Then
 
-// MARK: - UserDefaults에 사용되는 key 값 모음
-
-enum UserDefaultsKey: String {
-    case couchMark = "showCouchMark"
-}
-
 protocol PresentDelegate: AnyObject {
     func present(_ viewController: UIViewController)
 }
@@ -121,7 +115,7 @@ class DetailParentViewController: BaseViewController {
     }()
     // 세부 목표를 추가해주세요! 데이터
     private var emptyGoal: DetailGoal?
-    private var couchMarkKey: String = UserDefaultsKey.couchMark.rawValue
+    private var couchMarkKey: String = UserDefaultsKeyStyle.couchMark.rawValue
     
     // MARK: - Life Cycle
